🐛 : 버그 수정
Former-commit-id: 9313ddd8176df736d427fc2421534fa5f8aa7861

+++ backend/src/main/java/com/yam/funteer/user/service/LoginServiceImpl.java
@@ -65,8 +65,12 @@ public class LoginServiceImpl implements LoginService{
 
     @Override
     public void processLogOut() {
-        Long userId = SecurityUtil.getCurrentUserId();
-        tokenRepository.deleteById(userId);
+        if(SecurityUtil.isLogin()) {
+            Long userId = SecurityUtil.getCurrentUserId();
+            tokenRepository.deleteById(userId);
+        }
+
+        throw new AccessDeniedException("이미 로그아웃한 회원입니다.");
     }
 
     @Override
